Fix issue with stale toggle state, typo in modal blade
If the user changed the toggle and then in the "unsaved changes" dialogue clicked "discard," the toggle didn't refresh. It was due to a stale 'checked' attribute fighting with Livewire's internal state.

I initially added the 'checked' attribute so that the toggle would be correct on initial page load. But it turns out if I make sure the values of `$form` have the model's casts properly applied then that's not an issue. `->toArray()` applies the casts whereas `->getAttributes()` does not.