Fix console error when deleting new glyph after edit
Steps to reproduce:
1. Create a new glyph
2. Edit the glyph and save
3. Delete the glyph

Sometimes, this caused an error inside `$store.reorderFunctions.isFirstItem()` and `$store.reorderFunctions.isLastItem()` because they were being called on the dead glyph. This happened because somehow a weird Livewire/Alpine desync allowed the DOM element to persist and its `x-bind:disabled` attribute to execute after the backed object was gone.

Deleting the DOM element optimistically avoids the problem, plus adding a graceful exit if the reorder functions are called with bad input.